Cant remember what i was working on seems to Dodgeball game, but does not really look to be done. will commit as it does not interfer with other work to keep this work in progress.

+++ GameMaster/DataAccessLayer/MoveDTO.cs
@@ -9,6 +9,6 @@ namespace GameMaster.DataAccessLayer
     public class MoveDTO
     {
         public string Player { get; set; }
-        public List<int> Commands { get; set; }
+        public List<float> Commands { get; set; }
     }
 }
